refactor partially complete
