+++ gender_history/utilities.py
@@ -1,0 +1,4 @@
+from pathlib import Path
+import os
+
+BASE_PATH = Path(os.path.dirname(os.path.realpath(__file__))).parent
